avformat/movenc: Adjust baseMediaDecodeTime in tfdt with source DTS
baseMediaDecodeTime must contain the duration of all previous fragments.
Usually, ffmpeg starts counting at zero. Chrome/Firefox use this value
to order the segments, if it's not correct playout will stutter and
freeze.

To fix this, we simply use the start_dts of the muxed source, which
seems to work.

This fix is not ready for inclusion in ffmpeg mainline as our case of
producing segments that assume that there are segments that were encoded
before it (with a total duration of baseMediaDecodeTime) is a pretty
special one. However, specifying a segment offset (currently only used
for the filename AFAIK) also is, so who knows...